refactor(frontend): extract chat workspace data-fetching into focused hooks
